chore(kernel): split wasm and riscv kernel entrypoints
