+++ crates/jstz_kernel/Cargo.toml
@@ -16,7 +16,7 @@ crate-type = ["cdylib", "lib"]
 [dependencies]
 bincode.workspace = true
 hex.workspace = true
-futures.workspace = true
+tokio.workspace = true
 jstz_core = { path = "../jstz_core" }
 jstz_crypto = { path = "../jstz_crypto" }
 jstz_proto = { path = "../jstz_proto" }
@@ -24,7 +24,7 @@ num-traits.workspace = true
 serde.workspace = true
 tezos-smart-rollup.workspace = true
 tezos_crypto_rs.workspace = true
-tezos_data_encoding = "0.6.0"
+tezos_data_encoding.workspace = true
 
 [dev-dependencies]
 jstz_mock = { path = "../jstz_mock" }
@@ -32,3 +32,4 @@ jstz_utils = { path = "../jstz_utils" }
 
 [features]
 v2_runtime = ["jstz_proto/v2_runtime", "jstz_proto/kernel"]
+riscv_kernel = ["v2_runtime"]
