+++ crates/jstz_kernel/src/riscv_kernel.rs
@@ -1,0 +1,5 @@
+use tezos_smart_rollup::prelude::Runtime;
+
+pub async fn run(rt: &mut impl Runtime) {
+    todo!()
+}
